vty: Fixup sccp/ss7 configuration
The sccp/ss7 configuration is now fixed. The cs7 instance id is
implicitly detected from the bsc_addr or the msc_addr. Depending
on what is listed last. (I am not sure if that is wise, maybe
we should only use the local/bsc address to do the lookup).

Remove cs7-instance vty command

Modify VTY commands, so that the fixed API is used

Set msc->a.cs7_instance from the VTY to when msc/bsc addr is parsed

Fix the initalization to use osmo_sccp_simple_client_on_ss7_id() and
pass the cs7-instance id we determined from the vty.

The whole thing is not waterproof yet. We are still not at the
point where we allow to leave the local address out. This would
be fine, but when it is left out, the only way to determine the
cs7 instance is to use the msc_addr then.

We also might want to make sure to reach a state where all cs7
related config may be left out (like with the MSC)