increase RAN timeout in MGCP FSM
The MGCP FSM implements a timeout when waiting for the RAN to complete
the call (assignment complete, alerting, connect...). This timeout
is currently set to 10sec. This means if the other end did not pick
up after 10sec, the MGCP connection will be lost while the phone keeps
ringing. When the other end finally picks up, the call gets
disconnected.

This behavior is odd and requires a proper fix. For now increasing the
timeout to 120sec. will decrese the probability that he problem occurs.

- Increas RAN timeout to 120sec (2 min).

Change-Id: I5a11d53f9701d9b11b18d7026ff2241c7c0b57f5