msc_vlr_tests: revert IMSI parameter and test nr output
Three recently merged commits take the msc_vlr_tests in a wrong direction.

The IMSI is usually encoded in the hex streams. The rationale behind hex
streams is that it is a) easily copied from a wireshark trace and b) exactly
the bytes as sent by an actual phone. It is hard to parameterize the IMSI
because we would have to employ our encoding functions, which I intentionally
want to keep out of the loop here.

The test number should not appear in the normal test output, so that adding a
test or changing their order does not affect expected output for following
tests. The nr is simply for manual invocation, only seen when invoked with -v.

Revert
- "VLR tests: always print test parameters"
  b0a4314911140b1599cccfc8171fcdab4cd9bfab.
- "Expand VLR tests"
  d5feadeee8dd24f991df2892d6bcf0be8b0cf707.
- "Move IMSI into test parameters"
  093300d141c300651954473d73138b72de04d931.

Change-Id: Ie1b49237746751021da88f6f07bbb9f780d077c9