doc: update mgw settings in example config file
This example configuration files lack port settings for the mgcp
client. Lets explicitly assign a port for the MGW and a local port.
For the local port lets use the IETF port number + 1. The reason
for this is that the default config for osmo-bsc already uses the
IETF port and in osmo-bsc and osmo-msc run on the same machine in
many setups.

Change-Id: I17453e0d30eec757aba9530b63eb5d1539cbdffc