mncc: fix byte ordering of IP-Address in mncc
At the moment osmo-msc populates the member ip in struct gsm_mncc_rtp
with the wrong byte ordering. This causes LCR or
osmo-sip-connector to receive the IP address in the wrong order, which
eventually leads into a reversed IP address in the SDP part of the SIP
messages.

Change-Id: I86148179b549b511528e4c65213eb6c204cc609e
Related: OS#3431