msc_mgcp: move mncc struct initalization to where its actually needed
The function _handle_error() initalizes a struct gsm_mncc variable
on startup. The initalization accesses mgcp_ctx->trans->callref. All
this is done before the assertion on mgcp_ctx. Later in the code one
finds an if which tests on mgcp_ctx->free_ctx. This is the only part of
the code that accesses the mncc struct variable. We should move the
initalization there as well.

- Move initalization of struct gsm_mncc mncc into the if body
  that uses it.

Change-Id: I86983eabd999c4275dcc0e4a169ef2aa1e33c747
Related: OS#3635