libmsc: Don't send SMS STATUS REPORT locally if the ESME accepted it.
When using smpp-first, after the ESME accepts our STATUS REPORT,
we were sending it locally into gsm340_rx_sms_submit() anyway.
In the case of the ESME mirroring the report back to us, this
would result in two copies of the status report in the SMS
database, which were also both then delivered to the MS.
This causes no visible error to the user but is a waste of radio
resources.

With this patch, we check if it is the sms_report that has had 
receiver set in sms_route_mt_sms() and not the original SMS we
are reporting on, which of course already has receiver set.

Change-Id: I3529b89535800eaa1127721d613fa7bbcb8b23be