make: prepare for adding transcript tests
Separate 'make python-test' into separate make targets, to sensibly add VTY
transcript tests in an upcoming commit.

Feature: even though ./configure was called without --enable-external-tests,
each of the {ctrl,vty}x{python,transcript} tests can be invoked individually by
e.g. 'make vty-python-test'.

A peculiarity is that we use AM_PYTHON_PATH to figure out the 'python' command
to invoke as $(PYTHON), but it is empty when not configured with
--enable-external-tests. In that case, explain to the user that passing
'PYTHON=python' as make argument still allows running the tests. We could
always invoke AM_PYTHON_PATH, but that's not really needed.

Both 'vty-transcript-test' and 'ctrl-transcript-test' are still empty, a
subsequent patch adds a vty-transcript-test.

All of this in preparation of tweaking the 'mncc' vty configuration, to be able
to track it in a vty transcript test.

Change-Id: I688657e56ae469c07b9f25ba37275d38dbd457e2