msc_vty: display subscriber flags
When a VLR subscriber is displayed on the VTY we get a lot of meta
information, but there are also some flags to handle the internal
subscriber status e.g. conf_by_radio_contact_ind. Lets display those
flags as well as this information can be very helpful when debugging
problems in the VLR

Change-Id: I59a9145a4daad50d68de3fd5c3291f027256917f